Account System Loading
I add this system for multiple accounts including users and administrators. I have provided a way to register account. Of course, adminitrator accout need to have some secret code to identify the user in order to create the account.

PS: I hide the way to get all the accounts' user name and password.